LoraInterferenceHelper: fix u8 print

+++ model/lora-interference-helper.cc
@@ -330,7 +330,7 @@ LoraInterferenceHelper::IsDestroyedByInterference(Ptr<LoraInterferenceHelper::Ev
         if (snir >= snirIsolation)
         {
             // Move on and check the rest of the interferers
-            NS_LOG_DEBUG("Packet survived interference with SF " << currentSf);
+            NS_LOG_DEBUG("Packet survived interference with SF " << unsigned(currentSf));
         }
         else
         {
